feat(GUI): Fixed merge conflicts
