+++ BigBaseV2/src/services/gui_service.hpp
@@ -6,6 +6,8 @@ namespace big
 	enum class tabs {
 		ESP_SETTINGS,
 		GUI_SETTINGS,
+		NOTIFICATION_SETTINGS,
+		PROTECTION_SETTINGS,
 		DEBUG,
 		MOBILE,
 		NONE,
@@ -49,8 +51,10 @@ namespace big
 				{ tabs::SESSION, { "Session", view::session }},
 			}}},
 			{tabs::SETTINGS, { "Settings", view::settings, {
-				{ tabs::GUI_SETTINGS, { "GUI", view::gui_settings}},
 				{ tabs::ESP_SETTINGS, { "ESP", view::esp_settings}},
+				{ tabs::GUI_SETTINGS, { "GUI", view::gui_settings}},
+				{ tabs::NOTIFICATION_SETTINGS, { "Notifications", view::notification_settings}},
+				{ tabs::PROTECTION_SETTINGS, { "Protection", view::protection_settings}},
 				{ tabs::DEBUG, { "Debug", view::debug }},
 			}}},
 			{tabs::PLAYER, {"", view::view_player}}
