+++ BigBaseV2/src/views/core/view_active_view.cpp
