+++ BigBaseV2/src/views/core/view_heading.cpp
