+++ BigBaseV2/src/views/network/view_session.cpp
