+++ BigBaseV2/src/views/players/view_player.cpp
