+++ BigBaseV2/src/views/self/view_mobile.cpp
