+++ BigBaseV2/src/views/self/view_self.cpp
