+++ BigBaseV2/src/views/self/view_weapons.cpp
